+++ sdk/core/azure-core-tracing-opentelemetry/src/main/java/com/azure/core/tracing/opentelemetry/OpenTelemetryHttpPolicy.java
@@ -12,7 +12,6 @@ import com.azure.core.http.policy.AfterRetryPolicyProvider;
 import com.azure.core.http.policy.HttpPipelinePolicy;
 import com.azure.core.tracing.opentelemetry.implementation.HttpTraceUtil;
 import com.azure.core.util.CoreUtils;
-import com.azure.core.util.UrlBuilder;
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.trace.Span;
@@ -73,7 +72,12 @@ public class OpenTelemetryHttpPolicy implements AfterRetryPolicyProvider, HttpPi
     private static final String HTTP_METHOD = "http.method";
     private static final String HTTP_URL = "http.url";
     private static final String HTTP_STATUS_CODE = "http.status_code";
-    private static final String REQUEST_ID = "x-ms-request-id";
+    private static final String SERVICE_REQUEST_ID_HEADER = "x-ms-request-id";
+    private static final String SERVICE_REQUEST_ID_ATTRIBUTE = "serviceRequestId";
+
+    private static final String CLIENT_REQUEST_ID_HEADER = "x-ms-client-request-id";
+    private static final String CLIENT_REQUEST_ID_ATTRIBUTE = "requestId";
+
 
     // This helper class implements W3C distributed tracing protocol and injects SpanContext into the outgoing http
     // request
@@ -89,10 +93,12 @@ public class OpenTelemetryHttpPolicy implements AfterRetryPolicyProvider, HttpPi
         Span parentSpan = (Span) context.getData(PARENT_SPAN_KEY).orElse(Span.current());
         HttpRequest request = context.getHttpRequest();
 
-        // Build new child span representing this outgoing request.
-        final UrlBuilder urlBuilder = UrlBuilder.parse(context.getHttpRequest().getUrl());
-
-        SpanBuilder spanBuilder = tracer.spanBuilder(urlBuilder.getPath())
+        // Build new child span representing this outgoing request
+        // provide sampling-relevant attributes (users make sampling decisions based on this)
+        String methodName = request.getHttpMethod().toString();
+        SpanBuilder spanBuilder = tracer.spanBuilder("HTTP " + methodName)
+            .setAttribute(HTTP_METHOD, methodName)
+            .setAttribute(HTTP_URL, request.getUrl().toString())
             .setParent(currentContext.with(parentSpan));
 
         // A span's kind can be SERVER (incoming request) or CLIENT (outgoing request);
@@ -101,9 +107,9 @@ public class OpenTelemetryHttpPolicy implements AfterRetryPolicyProvider, HttpPi
         // Starting the span makes the sampling decision (nothing is logged at this time)
         Span span = spanBuilder.startSpan();
 
-        // If span is sampled in, add additional TRACING attributes
+        // If span is sampled in, add additional attributes
         if (span.isRecording()) {
-            addSpanRequestAttributes(span, request, context); // Adds HTTP method, URL, & user-agent
+            addPostSamplingAttributes(span, request, context);
         }
 
         // For no-op tracer, SpanContext is INVALID; inject valid span headers onto outgoing request
@@ -118,15 +124,16 @@ public class OpenTelemetryHttpPolicy implements AfterRetryPolicyProvider, HttpPi
             .contextWrite(Context.of("TRACING_SPAN", span, "REQUEST", request));
     }
 
-    private static void addSpanRequestAttributes(Span span, HttpRequest request,
+    private static void addPostSamplingAttributes(Span span, HttpRequest request,
         HttpPipelineCallContext context) {
         putAttributeIfNotEmptyOrNull(span, HTTP_USER_AGENT,
             request.getHeaders().getValue("User-Agent"));
-        putAttributeIfNotEmptyOrNull(span, HTTP_METHOD, request.getHttpMethod().toString());
-        putAttributeIfNotEmptyOrNull(span, HTTP_URL, request.getUrl().toString());
         Optional<Object> tracingNamespace = context.getData(AZ_TRACING_NAMESPACE_KEY);
         tracingNamespace.ifPresent(o -> putAttributeIfNotEmptyOrNull(span, OpenTelemetryTracer.AZ_NAMESPACE_KEY,
             o.toString()));
+
+        String requestId = request.getHeaders().getValue(CLIENT_REQUEST_ID_HEADER);
+        putAttributeIfNotEmptyOrNull(span, CLIENT_REQUEST_ID_ATTRIBUTE, requestId);
     }
 
     private static void putAttributeIfNotEmptyOrNull(Span span, String key, String value) {
@@ -183,10 +190,10 @@ public class OpenTelemetryHttpPolicy implements AfterRetryPolicyProvider, HttpPi
             String requestId = null;
             if (response != null) {
                 statusCode = response.getStatusCode();
-                requestId = response.getHeaderValue(REQUEST_ID);
+                requestId = response.getHeaderValue(SERVICE_REQUEST_ID_HEADER);
             }
 
-            putAttributeIfNotEmptyOrNull(span, REQUEST_ID, requestId);
+            putAttributeIfNotEmptyOrNull(span, SERVICE_REQUEST_ID_ATTRIBUTE, requestId);
             span.setAttribute(HTTP_STATUS_CODE, statusCode);
             span = HttpTraceUtil.setSpanStatus(span, statusCode, error);
         }